requirements.txt
langchain>=1.0.2
langchain-openai>=1.0.0
langchain-google-genai>=1.0.0
langchain-anthropic>=1.0.0
langchain-xai>=1.0.0
google-genai>=0.2.0
langgraph>=1.0.1
langgraph-checkpoint-sqlite>=1.0.0
python-dotenv>=1.2.1
chromadb>=0.4.0
langchain-chroma>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.2.0
langchain-huggingface>=0.0.3
sentence-transformers>=2.2.0
ddgs>=0.1.0
pymatgen>=2025.10.7
mp-api>=0.45.13
ase>=3.26.0
mace-torch>=0.3.14
lxml>=6.0.2
numpy>=1.26.0,<2.0
beautifulsoup4>=4.12.0
requests>=2.31.0
matplotlib
pandas
seaborn
pypdf>=4.0.0
pysocks
psutil